+++ neo/d3xp/Game_local.cpp
@@ -1259,6 +1259,71 @@ void idGameLocal::MapPopulate()
 
 /*
 ===================
+RB idGameLocal::PopulateEnvironmentProbes
+===================
+*/
+void idGameLocal::PopulateEnvironmentProbes()
+{
+	idEntity* ent;
+
+	// check if there are already environment probes defined by the artist
+	int numEnvprobes = 0;
+
+	for( ent = spawnedEntities.Next(); ent != NULL; ent = ent->spawnNode.Next() )
+	{
+		if( !ent->IsType( EnvironmentProbe::Type ) )
+		{
+			continue;
+		}
+
+		numEnvprobes++;
+	}
+
+	if( numEnvprobes > 0 )
+	{
+		return;
+	}
+
+	const idDict* envProbeDef = gameLocal.FindEntityDefDict( "env_probe", false );
+	if( !envProbeDef )
+	{
+		return;
+	}
+
+	// naive approach: place an env probe into the center of each BSP area
+
+	int	numAreas = gameRenderWorld->NumAreas();
+
+	for( int i = 0 ; i < numAreas ; i++ )
+	{
+		idBounds areaBounds = gameRenderWorld->AreaBounds( i );
+
+		idVec3 point = areaBounds.GetCenter();
+
+		int areaNum = gameRenderWorld->PointInArea( point );
+		if( areaNum < 0 )
+		{
+			Warning( "PopulateEnvironmentProbes: location '%i' is not in a valid area\n", i );
+			continue;
+		}
+
+		idDict args;
+		args.Set( "classname", "env_probe" );
+		args.Set( "origin", point.ToString() );
+
+		gameLocal.SpawnEntityDef( args, &ent );
+		if( !ent )
+		{
+			gameLocal.Error( "Couldn't spawn 'env_probe'" );
+		}
+
+		//environmentProbes.Append( probe );
+	}
+}
+// RB end
+
+/*
+===================
 idGameLocal::InitFromNewMap
 ===================
 */
@@ -1293,6 +1358,9 @@ void idGameLocal::InitFromNewMap( const char* mapName, idRenderWorld* renderWorl
 	InitScriptForMap();
 
 	MapPopulate();
+
+	// RB
+	PopulateEnvironmentProbes();
 
 	mpGame.Reset();
 	mpGame.Precache();
@@ -1564,7 +1632,6 @@ bool idGameLocal::InitFromSaveGame( const char* mapName, idRenderWorld* renderWo
 	savegame.RestoreObjects();
 
 	mpGame.Reset();
-
 	mpGame.Precache();
 
 	// free up any unused animations
@@ -3989,7 +4056,6 @@ void idGameLocal::SpawnMapEntities()
 	for( i = 1 ; i < numEntities ; i++ )
 	{
 		common->UpdateLevelLoadPacifier();
-
 
 		mapEnt = mapFile->GetEntity( i );
 		args = mapEnt->epairs;