+++ neo/renderer/OpenGL/RenderDebug_GL.cpp
@@ -1714,7 +1714,7 @@ void idRenderBackend::DBG_ShowViewEnvprobes()
 
 	GL_State( GLS_DEFAULT | GLS_CULL_TWOSIDED );
 
-	renderProgManager.BindShader_Environment();
+	renderProgManager.BindShader_Octahedron();
 
 	int count = 0;
 	for( viewEnvprobe_t* vProbe = viewDef->viewEnvprobes; vProbe != NULL; vProbe = vProbe->next )