+++ neo/renderer/RenderBackend.cpp
@@ -6003,7 +6003,7 @@ void idRenderBackend::PostProcess( const void* data )
 {
 	// only do the post process step if resolution scaling is enabled. Prevents the unnecessary copying of the framebuffer and
 	// corresponding full screen quad pass.
-	if( rs_enable.GetInteger() == 0 && !r_useFilmicPostProcessEffects.GetBool() && r_antiAliasing.GetInteger() == 0 )
+	if( rs_enable.GetInteger() == 0 && !r_useFilmicPostProcessing.GetBool() && r_antiAliasing.GetInteger() == 0 )
 	{
 		return;
 	}
@@ -6113,7 +6113,7 @@ void idRenderBackend::PostProcess( const void* data )
 #endif
 	}
 
-	if( r_useFilmicPostProcessEffects.GetBool() )
+	if( r_useFilmicPostProcessing.GetBool() )
 	{
 		globalImages->currentRenderImage->CopyFramebuffer( viewport.x1, viewport.y1, viewport.GetWidth(), viewport.GetHeight() );