+++ neo/renderer/RenderWorld_load.cpp
@@ -1090,7 +1090,6 @@ void idRenderWorldLocal::AddWorldModelEntities()
 	{
 		common->UpdateLevelLoadPacifier();
 
-
 		idRenderEntityLocal*	 def = new( TAG_RENDER_ENTITY ) idRenderEntityLocal;
 
 		// try and reuse a free spot
@@ -1142,7 +1141,11 @@ void idRenderWorldLocal::AddWorldModelEntities()
 
 		R_DeriveEntityData( def );
 
-		AddEntityRefToArea( def, &portalAreas[i] );
+		portalArea_t* area = &portalAreas[i];
+		AddEntityRefToArea( def, area );
+
+		// RB: remember BSP area AABB for quick lookup later
+		area->globalBounds = def->globalReferenceBounds;
 	}
 }